feat(cmd,eth): accept hex and relative offsets in --set-head
The flag was a cli.Uint64Flag, so it went through Go's base-0 integer
parsing: "--set-head=010" silently meant 8 (octal), "--set-head=0b101"
meant 5, and a negative value was rejected outright, leaving operators to
compute the target block number by hand while recovering a node.

Parse the raw string ourselves instead: "0x"-prefixed values are read as
hexadecimal, everything else as decimal, and a negative value is the number
of blocks to roll back from the current head (e.g. --set-head=-1000).
Octal and binary literals, "_" separators, leading zeroes and a "+" sign
are rejected on purpose, so that every accepted input has exactly one
reading. The value is stored in common.RollbackNumber as int64, and the
offset is resolved against the current block once the chain is open;
asking to roll back more blocks than the chain is high is an error.

The old "the flag --set-head must be greater than 0" message no longer
holds now that negative values are valid, and is replaced.

Note for operators: "--set-head=18446744073709551615" used to be accepted
and is now out of range for an int64 block number.

Tests: cmd/utils/flags_sethead_test.go covers the accepted forms (decimal,
0x hex, negative offset, negative hex) and the rejected ones (octal and
binary literals, "_" separators, leading zeroes, "+" sign, multiple signs,
values beyond the int64 range).

+++ core/blockchain_test.go
@@ -605,7 +605,7 @@ func TestNewBlockChainRepairsMissingHeadStateRespectsRollbackTarget(t *testing.T
 	rawdb.DeleteLegacyTrieNode(db, head.Root())
 
 	prevRollback := common.RollbackNumber
-	common.RollbackNumber = target.NumberU64()
+	common.RollbackNumber = int64(target.NumberU64())
 	t.Cleanup(func() {
 		common.RollbackNumber = prevRollback
 	})
